requirements.txt
pandas
matplotlib
seaborn
scikit-learn
sentence-transformers
faiss-cpu         # or faiss-gpu if you have CUDA
langchain
openai
streamlit
gradio
tqdm
streamlit
faiss-cpu
sentence-transformers
transformers
torch
